fix some stuff that generated warnings

+++ src/util/set_minus.h
@@ -14,10 +14,10 @@
  * @return
  */
 template<typename T, typename Set, typename Minus, typename Out>
-int set_minus_sorted(const Set &set, const Minus &minus, Out &out) {
-    int i = 0;
-    int j = 0;
-    int k = 0;
+size_t set_minus_sorted(const Set &set, const Minus &minus, Out &out) {
+    size_t i = 0;
+    size_t j = 0;
+    size_t k = 0;
 
     while (i < set.size()) {
         if (i > 0 && set[i] < set[i - 1]) break;
@@ -35,10 +35,10 @@ int set_minus_sorted(const Set &set, const Minus &minus, Out &out) {
 };
 
 template<typename T>
-int set_minus_sorted(const std::vector<T> &set, const std::vector<T> &minus, std::vector<T> &out) {
+size_t set_minus_sorted(const std::vector<T> &set, const std::vector<T> &minus, std::vector<T> &out) {
     std::span<const T> span1(set.begin(), set.end());
     std::span<const T> span2(minus.begin(), minus.end());
-    int k = set_minus_sorted<T, std::span<const T>, std::span<const T>, std::vector<T>>(span1, span2, out);
+    size_t k = set_minus_sorted<T, std::span<const T>, std::span<const T>, std::vector<T>>(span1, span2, out);
     out.resize(k);
     out.shrink_to_fit();
